Minor display bugfixes and Using distinct icons for File and Folder nodes is required.

+++ WaveManagerBusiness/WaveManager.cs
@@ -38,7 +38,10 @@ namespace WaveManagerBusiness
         }
         private static void RemoveOpenFile(WaveFile file)
         {
-            WaveFileRepository.RemoveFile(file.filePath);
+            if (file != null)
+            {
+                WaveFileRepository.RemoveFile(file.filePath);
+            }
             ActiveFile = null;
         }
